users/requirements.txt
amqp==5.3.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.3
greenlet==3.2.2
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
prompt_toolkit==3.0.51
protobuf==5.29.4
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.3.0
setuptools==80.7.1
six==1.17.0
SQLAlchemy==2.0.41
tenacity==8.2.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
